refactor(ui): move existing components to panda style folders
