+++ packages/ui/src/component/accordion/panda/recipe.ts
