+++ packages/ui/src/component/animateText/panda/index.tsx
