+++ packages/ui/src/component/avatar/panda/index.tsx
