+++ packages/ui/src/component/avatar/panda/recipe.ts
