+++ packages/ui/src/component/checkbox/panda/recipe.ts
