+++ packages/ui/src/component/slider/panda/index.tsx
