+++ packages/ui/src/component/slider/panda/recipe.ts
